@@ -21,16 +21,18 @@ android {
         }
     }
 
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
 
     defaultConfig {
         applicationId "de.stephanlindauer.criticalmaps"
         minSdkVersion 16
-        targetSdkVersion 25
-        versionCode 34
-        versionName "2.4.2"
+        targetSdkVersion 28
+        versionCode 35
+        versionName "2.5.0"
         vectorDrawables.useSupportLibrary = true
+
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
 
         javaCompileOptions {
             annotationProcessorOptions {
@@ -44,6 +46,7 @@ android {
     buildTypes {
         debug {
             applicationIdSuffix ".debug"
+            pseudoLocalesEnabled true
         }
         release {
             minifyEnabled false
@@ -55,6 +58,7 @@ android {
     lintOptions {
         warning 'MissingTranslation'
         warning 'StringFormatInvalid'
+        warning 'NewApi'
 
         // https://github.com/square/okio/issues/58
         // http://jakewharton.github.io/butterknife/
@@ -73,29 +77,40 @@ android {
 
 dependencies {
     implementation 'com.squareup:otto:1.3.8'
-    implementation 'org.osmdroid:osmdroid-android:6.0.2-SNAPSHOT'
+    implementation 'org.osmdroid:osmdroid-android:6.1.0'
     implementation 'com.squareup.picasso:picasso:2.5.2'
     implementation 'com.jakewharton.picasso:picasso2-okhttp3-downloader:1.1.0'
-    implementation "com.android.support:support-v4:$support_version"
-    implementation "com.android.support:appcompat-v7:$support_version"
-    implementation "com.android.support:support-annotations:$support_version"
-    implementation "com.android.support:design:$support_version"
-    implementation "com.android.support:exifinterface:$support_version"
-    implementation 'com.jakewharton.timber:timber:4.7.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.appcompat:appcompat:1.0.2'
+    implementation 'androidx.annotation:annotation:1.0.2'
+    implementation 'com.google.android.material:material:1.1.0-alpha05'
+    implementation 'androidx.exifinterface:exifinterface:1.0.0'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+    implementation 'com.jakewharton.timber:timber:4.7.1'
 
-    implementation 'com.squareup.okhttp3:okhttp:3.10.0'
+    implementation 'com.squareup.okhttp3:okhttp:3.12.2'
     implementation "com.jakewharton:butterknife:$butterknife_version"
     annotationProcessor "com.jakewharton:butterknife-compiler:$butterknife_version"
 
     implementation 'org.ligi:AXT:0.37'
     implementation 'info.metadude.android:typed-preferences:2.1.0'
-    implementation "com.google.dagger:dagger:$dagger_version"
 
     compileOnly 'javax.annotation:jsr250-api:1.0'
+    implementation "com.google.dagger:dagger:$dagger_version"
     annotationProcessor "com.google.dagger:dagger-compiler:$dagger_version"
 
+    debugImplementation 'com.squareup.leakcanary:leakcanary-android:2.0-alpha-1'
+
+    errorprone("com.google.errorprone:error_prone_core:2.3.3")
+    errorproneJavac("com.google.errorprone:javac:9+181-r4173-1")
+
     testImplementation 'junit:junit:4.12'
-    testImplementation 'com.google.truth:truth:0.40'
-    testImplementation "com.android.support:support-annotations:$support_version"
-    testImplementation 'org.mockito:mockito-core:2.15.0'
+    testImplementation 'com.google.truth:truth:0.41'
+    testImplementation 'org.mockito:mockito-core:2.23.0'
+
+    androidTestImplementation 'androidx.test:core:1.1.0'
+    androidTestImplementation 'androidx.test:runner:1.1.1'
+    androidTestImplementation 'androidx.test:rules:1.1.1'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.0'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.1'
 }
