@@ -1,23 +1,42 @@
 package de.stephanlindauer.criticalmaps;
 
+import android.animation.ArgbEvaluator;
+import android.animation.ValueAnimator;
+import android.annotation.TargetApi;
 import android.app.Activity;
 import android.content.Intent;
+import android.content.SharedPreferences;
+import android.graphics.drawable.GradientDrawable;
+import android.net.Uri;
+import android.os.Build;
 import android.os.Bundle;
-import android.support.annotation.IdRes;
-import android.support.annotation.NonNull;
-import android.support.design.widget.NavigationView;
-import android.support.v4.app.Fragment;
-import android.support.v4.view.GravityCompat;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBarDrawerToggle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
+import androidx.annotation.IdRes;
+import androidx.annotation.NonNull;
+import com.google.android.material.navigation.NavigationView;
+
+import androidx.core.content.ContextCompat;
+import androidx.core.view.ViewCompat;
+import androidx.fragment.app.Fragment;
+import androidx.core.view.GravityCompat;
+import androidx.drawerlayout.widget.DrawerLayout;
+import androidx.appcompat.app.ActionBarDrawerToggle;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.SwitchCompat;
+import androidx.appcompat.widget.Toolbar;
+
 import android.util.SparseArray;
 import android.view.Menu;
 import android.view.MenuItem;
+import android.view.View;
+import android.view.ViewGroup;
 import android.view.WindowManager;
+import android.view.animation.AccelerateDecelerateInterpolator;
+import android.widget.FrameLayout;
 import android.widget.Toast;
 
+import org.jetbrains.annotations.NotNull;
+
+import androidx.recyclerview.widget.RecyclerView;
 import butterknife.BindView;
 import java.io.File;
 
@@ -31,25 +50,31 @@ import de.stephanlindauer.criticalmaps.handler.ProcessCameraResultHandler;
 import de.stephanlindauer.criticalmaps.handler.StartCameraHandler;
 import de.stephanlindauer.criticalmaps.helper.clientinfo.BuildInfo;
 import de.stephanlindauer.criticalmaps.helper.clientinfo.DeviceInformation;
+import de.stephanlindauer.criticalmaps.prefs.SharedPrefsKeys;
 import de.stephanlindauer.criticalmaps.provider.FragmentProvider;
 import de.stephanlindauer.criticalmaps.service.ServerSyncService;
 import de.stephanlindauer.criticalmaps.utils.DrawerClosingDrawerLayoutListener;
 import de.stephanlindauer.criticalmaps.utils.ImageUtils;
 import de.stephanlindauer.criticalmaps.utils.IntentUtil;
 import de.stephanlindauer.criticalmaps.vo.RequestCodes;
+import info.metadude.android.typedpreferences.BooleanPreference;
 import timber.log.Timber;
 
 public class Main extends AppCompatActivity implements NavigationView.OnNavigationItemSelectedListener {
 
     private final static String KEY_NAVID = "main_navid";
     private final static String KEY_SAVEDFRAGMENTSTATES = "main_savedfragmentstate";
+    private final static String KEY_NEWCAMERAOUTPUTFILE = "main_newcameraoutputfile";
 
-    private File newCameraOutputFile;
+    private Uri newCameraOutputFile;
     private int currentNavId;
     private SparseArray<Fragment.SavedState> savedFragmentStates = new SparseArray<>();
 
     @Inject
     public PermissionCheckHandler permissionCheckHandler;
+
+    @Inject
+    SharedPreferences sharedPreferences;
 
     @BindView(R.id.drawer_layout)
     DrawerLayout drawerLayout;
@@ -60,44 +85,65 @@ public class Main extends AppCompatActivity implements NavigationView.OnNavigati
     @BindView(R.id.toolbar)
     Toolbar toolbar;
 
-    @Override
-    protected void onPostCreate(Bundle savedInstanceState) {
-        super.onPostCreate(savedInstanceState);
+    @BindView(R.id.content_frame)
+    FrameLayout contentFrame;
 
-        drawerNavigation.setNavigationItemSelectedListener(this);
-
-        setSupportActionBar(toolbar);
-        ActionBarDrawerToggle mDrawerToggle = new ActionBarDrawerToggle(this, drawerLayout, toolbar, R.string.open_drawer, R.string.close_drawer);
-
-        drawerLayout.addDrawerListener(mDrawerToggle);
-        mDrawerToggle.syncState();
-        drawerLayout.addDrawerListener(new DrawerClosingDrawerLayoutListener());
-
-        if (savedInstanceState != null) {
-            SparseArray<Fragment.SavedState> restoredStates = savedInstanceState.getSparseParcelableArray(KEY_SAVEDFRAGMENTSTATES);
-            if (restoredStates != null) {
-                savedFragmentStates = restoredStates;
-            }
-
-            currentNavId = savedInstanceState.getInt(KEY_NAVID);
-        } else {
-            navigateTo(R.id.navigation_map);
-        }
-    }
+    private final SharedPreferences.OnSharedPreferenceChangeListener sharedPreferenceChangeListener =
+            (sharedPreferences, key) -> {
+                if (SharedPrefsKeys.SHOW_ON_LOCKSCREEN.equals(key)) {
+                    setShowOnLockscreen();
+                } else if(SharedPrefsKeys.KEEP_SCREEN_ON.equals(key)) {
+                    setKeepScreenOn();
+                }
+            };
 
     @Override
     public void onCreate(Bundle bundle) {
         setTheme(R.style.AppTheme); // has to be before super!
-
         super.onCreate(bundle);
 
         App.components().inject(this);
-
         setContentView(R.layout.activity_main);
-
         ButterKnife.bind(this);
 
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+            drawerLayout.setSystemUiVisibility(View.SYSTEM_UI_FLAG_LAYOUT_STABLE
+                    | View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN);
+
+            // inset the toolbar down by the status bar height
+            ViewCompat.setOnApplyWindowInsetsListener(toolbar, (v, insets) -> {
+                ViewGroup.MarginLayoutParams lpToolbar =
+                        (ViewGroup.MarginLayoutParams) toolbar.getLayoutParams();
+                lpToolbar.topMargin += insets.getSystemWindowInsetTop();
+
+                toolbar.setLayoutParams(lpToolbar);
+
+                // clear this listener so insets aren't re-applied
+                ViewCompat.setOnApplyWindowInsetsListener(toolbar, null);
+                return insets;
+            });
+
+            // inset header in nav drawer down by the status bar height
+            View navHeader = drawerNavigation.getHeaderView(0);
+            ViewCompat.setOnApplyWindowInsetsListener(navHeader, (v, insets) -> {
+                v.setPaddingRelative(
+                        v.getPaddingStart(), v.getPaddingTop() + insets.getSystemWindowInsetTop(),
+                        v.getPaddingEnd(), v.getPaddingBottom());
+
+                // clear this listener so insets aren't re-applied
+                ViewCompat.setOnApplyWindowInsetsListener(navHeader, null);
+                return insets;
+            });
+        }
+
+        // This is a little hacky and might break with a materialcomponents lib update
+        RecyclerView navigationMenuView = findViewById(R.id.design_navigation_view);
+        navigationMenuView.setNestedScrollingEnabled(false);
+
         new PrerequisitesChecker(this).showIntroductionIfNotShownBefore();
+
+        setShowOnLockscreen();
+        setKeepScreenOn();
 
         ServerSyncService.startService();
     }
@@ -106,17 +152,83 @@ public class Main extends AppCompatActivity implements NavigationView.OnNavigati
     protected void onStart() {
         super.onStart();
         permissionCheckHandler.attachActivity(this);
+        sharedPreferences.registerOnSharedPreferenceChangeListener(
+                sharedPreferenceChangeListener);
+    }
+
+    @Override
+    protected void onPostCreate(Bundle savedInstanceState) {
+        super.onPostCreate(savedInstanceState);
+
+        drawerNavigation.setNavigationItemSelectedListener(this);
+
+        setSupportActionBar(toolbar);
+        ActionBarDrawerToggle mDrawerToggle = new ActionBarDrawerToggle(
+                this, drawerLayout, toolbar, R.string.open_drawer, R.string.close_drawer);
+
+        drawerLayout.addDrawerListener(mDrawerToggle);
+        mDrawerToggle.syncState();
+        drawerLayout.addDrawerListener(new DrawerClosingDrawerLayoutListener());
+
+        SwitchCompat observerModeSwitch =
+                drawerNavigation.getMenu().findItem(R.id.navigation_observer_mode)
+                .getActionView().findViewById(R.id.navigation_observer_mode_switch);
+        observerModeSwitch.setChecked(new BooleanPreference(
+                sharedPreferences, SharedPrefsKeys.OBSERVER_MODE_ACTIVE).get());
+        observerModeSwitch.setOnCheckedChangeListener(
+                (buttonView, isChecked) -> handleObserverModeSwitchCheckedChanged(isChecked));
+
+        if (savedInstanceState != null) {
+            SparseArray<Fragment.SavedState> restoredStates =
+                    savedInstanceState.getSparseParcelableArray(KEY_SAVEDFRAGMENTSTATES);
+            if (restoredStates != null) {
+                savedFragmentStates = restoredStates;
+            }
+
+            newCameraOutputFile = savedInstanceState.getParcelable(KEY_NEWCAMERAOUTPUTFILE);
+
+            currentNavId = savedInstanceState.getInt(KEY_NAVID);
+            if (currentNavId != R.id.navigation_map) {
+                // set toolbar title
+                //noinspection ConstantConditions
+                getSupportActionBar().setTitle(drawerNavigation.getCheckedItem().getTitle());
+
+                // set toolbar margins
+                ViewGroup.MarginLayoutParams toolbarParams =
+                        (ViewGroup.MarginLayoutParams) toolbar.getLayoutParams();
+                int marginPixels =
+                        getResources().getDimensionPixelSize(R.dimen.map_toolbar_margins);
+
+                toolbarParams.topMargin -= marginPixels;
+                toolbarParams.rightMargin -= marginPixels;
+                toolbarParams.leftMargin -= marginPixels;
+                toolbar.setLayoutParams(toolbarParams);
+
+                // set toolbar background
+                ((GradientDrawable) toolbar.getBackground()).setCornerRadius(0F);
+
+                // set statusbar color
+                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+                    getWindow().setStatusBarColor(
+                            ContextCompat.getColor(this, R.color.main_statusbarcolor_others));
+                }
+            }
+        } else {
+            navigateTo(R.id.navigation_map);
+        }
     }
 
     @Override
     protected void onStop() {
         permissionCheckHandler.detachActivity();
+        sharedPreferences.unregisterOnSharedPreferenceChangeListener(
+                sharedPreferenceChangeListener);
         super.onStop();
     }
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
-        getMenuInflater().inflate(R.menu.actionbar_buttons, menu);
+        getMenuInflater().inflate(R.menu.menu_toolbar, menu);
         return true;
     }
 
@@ -151,18 +263,20 @@ public class Main extends AppCompatActivity implements NavigationView.OnNavigati
     private void startFeedbackIntent() {
         Intent Email = new Intent(Intent.ACTION_SEND);
         Email.setType("text/email");
-        Email.putExtra(Intent.EXTRA_EMAIL, new String[]{"stephanlindauer@posteo.de"});
+        Email.putExtra(Intent.EXTRA_EMAIL, new String[]{getString(R.string.contact_email)});
         Email.putExtra(Intent.EXTRA_SUBJECT, "feedback critical maps");
         Email.putExtra(Intent.EXTRA_TEXT, DeviceInformation.getString() + BuildInfo.getString());
         startActivity(Intent.createChooser(Email, "Send Feedback:"));
     }
 
     private void startDatenschutzIntent() {
-        IntentUtil.startFromURL(this, "http://criticalmaps.net/info#Datenschutzerklärung");
+        IntentUtil.startFromURL(this,
+                "http://criticalmaps.net/info#Datenschutzerklärung");
     }
 
     private void startRateTheApp() {
-        IntentUtil.startFromURL(this, "https://play.google.com/store/apps/details?id=de.stephanlindauer.criticalmaps");
+        IntentUtil.startFromURL(this,
+                "https://play.google.com/store/apps/details?id=de.stephanlindauer.criticalmaps");
     }
 
     @Override
@@ -176,14 +290,14 @@ public class Main extends AppCompatActivity implements NavigationView.OnNavigati
         }
 
         if (requestCode == RequestCodes.CAMERA_CAPTURE_IMAGE_REQUEST_CODE) {
-            File movedFile = ImageUtils.movePhotoToPublicDir(newCameraOutputFile);
+            File movedFile =
+                    ImageUtils.movePhotoToPublicDir(new File(newCameraOutputFile.getPath()));
             newCameraOutputFile = null;
             new ProcessCameraResultHandler(this, movedFile).execute();
         }
     }
 
-    // TODO use URI instead of File and save it with state, otherwise rotation while loose the reference
-    public void setNewCameraOutputFile(File newCameraOutputFile) {
+    public void setNewCameraOutputFile(Uri newCameraOutputFile) {
         this.newCameraOutputFile = newCameraOutputFile;
     }
 
@@ -195,24 +309,44 @@ public class Main extends AppCompatActivity implements NavigationView.OnNavigati
         }
     }
 
-    @Override
-    public void onAttachedToWindow() {
-        getWindow().addFlags(WindowManager.LayoutParams.FLAG_SHOW_WHEN_LOCKED);
+    private void setShowOnLockscreen() {
+        if (new BooleanPreference(sharedPreferences, SharedPrefsKeys.SHOW_ON_LOCKSCREEN).get()) {
+            getWindow().addFlags(WindowManager.LayoutParams.FLAG_SHOW_WHEN_LOCKED);
+        } else {
+            getWindow().clearFlags(WindowManager.LayoutParams.FLAG_SHOW_WHEN_LOCKED);
+        }
+    }
+
+    private void setKeepScreenOn() {
+        if (new BooleanPreference(sharedPreferences, SharedPrefsKeys.KEEP_SCREEN_ON).get()) {
+            getWindow().addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
+        } else {
+            getWindow().clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
+        }
     }
 
     @Override
-    protected void onSaveInstanceState(Bundle outState) {
+    protected void onSaveInstanceState(@NotNull Bundle outState) {
         super.onSaveInstanceState(outState);
         outState.putInt(KEY_NAVID, currentNavId);
         outState.putSparseParcelableArray(KEY_SAVEDFRAGMENTSTATES, savedFragmentStates);
+        if (newCameraOutputFile != null) {
+            outState.putParcelable(KEY_NEWCAMERAOUTPUTFILE, newCameraOutputFile);
+        }
     }
 
     @Override
     public boolean onNavigationItemSelected(@NonNull MenuItem item) {
-        item.setChecked(true);
-        drawerLayout.closeDrawer(GravityCompat.START);
-        navigateTo(item.getItemId());
-        return true;
+        if(item.getGroupId() == R.id.navigation_group) {
+            item.setChecked(true);
+            drawerLayout.closeDrawer(GravityCompat.START);
+            //noinspection ConstantConditions
+            getSupportActionBar().setTitle(item.getTitle());
+            navigateTo(item.getItemId());
+            return true;
+        }
+        //TODO set switch checked when observer switch item selected?
+        return false;
     }
 
     private void navigateTo(@IdRes int navId) {
@@ -223,17 +357,94 @@ public class Main extends AppCompatActivity implements NavigationView.OnNavigati
         // save state of current fragment
         Fragment currentFragment = getSupportFragmentManager().findFragmentById(R.id.content_frame);
         if (currentFragment != null) {
-            Fragment.SavedState state = getSupportFragmentManager().saveFragmentInstanceState(currentFragment);
+            Fragment.SavedState state =
+                    getSupportFragmentManager().saveFragmentInstanceState(currentFragment);
             savedFragmentStates.put(currentNavId, state);
         }
 
-        currentNavId = navId;
         final Fragment nextFragment = FragmentProvider.getFragmentForNavId(navId);
 
         // restore saved state of new fragment if it was shown before; otherwise passing null is ok
         nextFragment.setInitialSavedState(savedFragmentStates.get(navId));
 
-        getSupportFragmentManager().beginTransaction().replace(R.id.content_frame, nextFragment).commit();
+        getSupportFragmentManager().beginTransaction()
+                .replace(R.id.content_frame, nextFragment).commit();
+
+        // animate toolbar and statusbar color
+        if (currentNavId == R.id.navigation_map) {
+            // from map to other fragment
+            animateToolbar(200, false);
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+                fadeInStatusBarColor(200, false);
+            }
+        } else if (navId == R.id.navigation_map && currentNavId != 0) {
+            // from other fragment to map AND not app start
+            animateToolbar(500, true);
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+                fadeInStatusBarColor(500, true);
+            }
+        }
+
+        currentNavId = navId;
+    }
+
+    private void animateToolbar(int durationMillis, boolean toMap) {
+        ViewGroup.MarginLayoutParams toolbarParamsChanging =
+                (ViewGroup.MarginLayoutParams) toolbar.getLayoutParams();
+
+        ViewGroup.MarginLayoutParams toolbarParamsStart =
+                new ViewGroup.MarginLayoutParams(toolbarParamsChanging);
+
+        int marginPixels = getResources().getDimensionPixelSize(R.dimen.map_toolbar_margins);
+
+        if (toMap) {
+            marginPixels *= -1;
+        }
+
+        ValueAnimator marginAnimator = ValueAnimator.ofInt(0, marginPixels);
+        marginAnimator.setDuration(durationMillis);
+        marginAnimator.setInterpolator(new AccelerateDecelerateInterpolator());
+        marginAnimator.addUpdateListener(animation -> {
+            int animatedValue = (int) animation.getAnimatedValue();
+            toolbarParamsChanging.topMargin = toolbarParamsStart.topMargin - animatedValue;
+            toolbarParamsChanging.rightMargin = toolbarParamsStart.rightMargin - animatedValue;
+            toolbarParamsChanging.leftMargin = toolbarParamsStart.leftMargin - animatedValue;
+            toolbar.setLayoutParams(toolbarParamsChanging);
+        });
+
+        GradientDrawable toolbarBackground = (GradientDrawable) toolbar.getBackground();
+        float radiusMap = getResources().getDimension(R.dimen.map_toolbar_corner_radius);
+        float radiusFrom = toMap ? 0 : radiusMap;
+        float radiusTo = toMap ? radiusMap : 0;
+
+        ValueAnimator radiusAnimator = ValueAnimator.ofFloat(radiusFrom, radiusTo);
+        radiusAnimator.setDuration(durationMillis);
+        marginAnimator.setInterpolator(new AccelerateDecelerateInterpolator());
+        radiusAnimator.addUpdateListener(animation -> {
+                    float value = (float) animation.getAnimatedValue();
+                    toolbarBackground.setCornerRadius(value);
+                });
+
+        marginAnimator.start();
+        radiusAnimator.start();
+    }
+
+    @TargetApi(Build.VERSION_CODES.LOLLIPOP)
+    private void fadeInStatusBarColor(int duration, boolean toMap) {
+        int colorMap = ContextCompat.getColor(this, R.color.main_statusbarcolor_map);
+        int colorOthers = ContextCompat.getColor(this, R.color.main_statusbarcolor_others);
+        int colorFrom = toMap ? colorOthers : colorMap;
+        int colorTo = toMap ? colorMap : colorOthers;
+
+        ValueAnimator valueAnimator =
+                ValueAnimator.ofObject(new ArgbEvaluator(), colorFrom, colorTo);
+        valueAnimator.setDuration(duration);
+        valueAnimator.addUpdateListener(animation -> {
+            int color = (int) animation.getAnimatedValue();
+            getWindow().setStatusBarColor(color);
+        });
+
+        valueAnimator.start();
     }
 
     @Override
@@ -242,5 +453,10 @@ public class Main extends AppCompatActivity implements NavigationView.OnNavigati
         if (!permissionCheckHandler.handlePermissionRequestCallback(requestCode, grantResults)) {
             super.onRequestPermissionsResult(requestCode, permissions, grantResults);
         }
+    }
+
+    private void handleObserverModeSwitchCheckedChanged(boolean isChecked) {
+        new BooleanPreference(
+                sharedPreferences, SharedPrefsKeys.OBSERVER_MODE_ACTIVE).set(isChecked);
     }
 }
