@@ -8,8 +8,8 @@ import android.net.Uri;
 import android.os.Build;
 import android.provider.MediaStore;
 import android.provider.Settings;
-import android.support.v4.content.FileProvider;
-import android.support.v7.app.AlertDialog;
+import androidx.core.content.FileProvider;
+import androidx.appcompat.app.AlertDialog;
 
 import de.stephanlindauer.criticalmaps.App;
 import de.stephanlindauer.criticalmaps.Main;
@@ -64,7 +64,7 @@ public class StartCameraHandler {
             AlertBuilder.show(activity, R.string.something_went_wrong, R.string.camera_no_outputfile);
             return;
         }
-        activity.setNewCameraOutputFile(outputFile); // FIXME: 07.09.2015
+        activity.setNewCameraOutputFile(Uri.fromFile(outputFile));
 
         Uri imageCaptureUri = FileProvider.getUriForFile(
                 activity,
