@@ -1,14 +1,7 @@
 package de.stephanlindauer.criticalmaps.utils;
 
 import android.app.Activity;
-import android.content.Context;
-import android.content.SharedPreferences;
-import android.os.Build;
-import android.os.Environment;
-import android.os.StatFs;
-import android.support.annotation.NonNull;
-import android.support.v4.content.ContextCompat;
-import android.support.v4.os.EnvironmentCompat;
+import androidx.core.content.ContextCompat;
 import android.view.ViewGroup;
 
 import org.osmdroid.config.Configuration;
@@ -18,39 +11,37 @@ import org.osmdroid.tileprovider.modules.SqlTileWriter;
 import org.osmdroid.tileprovider.tilesource.OnlineTileSourceBase;
 import org.osmdroid.tileprovider.tilesource.XYTileSource;
 import org.osmdroid.util.GeoPoint;
+import org.osmdroid.views.CustomZoomButtonsController;
 import org.osmdroid.views.MapView;
 
 import java.io.File;
-import java.io.IOException;
-import java.util.ArrayList;
 
 import de.stephanlindauer.criticalmaps.App;
 import de.stephanlindauer.criticalmaps.BuildConfig;
 import de.stephanlindauer.criticalmaps.R;
-import de.stephanlindauer.criticalmaps.prefs.SharedPrefsKeys;
-import de.stephanlindauer.criticalmaps.provider.MapConfigurationProvider;
-import info.metadude.android.typedpreferences.StringPreference;
+import de.stephanlindauer.criticalmaps.provider.StorageLocationProvider;
 import timber.log.Timber;
 
 public class MapViewUtils {
+    
+    private final static OnlineTileSourceBase WIKIMEDIA = new XYTileSource("Wikimedia",
+            1, 19, 256, ".png",
+            new String[] {"https://maps.wikimedia.org/osm-intl/"},
+            "Wikimedia maps | Map data © OpenStreetMap contributors");
+
     private MapViewUtils() {}
 
     public static MapView createMapView(Activity activity) {
-        Configuration.setConfigurationProvider(new MapConfigurationProvider());
         IConfigurationProvider configuration = Configuration.getInstance();
 
-        SharedPreferences sharedPrefs = App.components().sharedPreferences();
-        StringPreference osmdroidBasePathPref =
-                new StringPreference(sharedPrefs, SharedPrefsKeys.OSMDROID_BASE_PATH);
+        StorageLocationProvider.StorageLocation storageLocation =
+                App.components().storageProvider().getActiveStorageLocation();
+        if (storageLocation == null) {
+            storageLocation = App.components().storageProvider().getAndSaveBestStorageLocation();
+        }
+        File osmdroidBasePath = storageLocation.osmdroidBasePath;
+        File osmdroidTileCache = storageLocation.osmdroidTilePath;
 
-        File osmdroidBasePath =
-                checkAndGetOsmdroidBasePathFile(activity, osmdroidBasePathPref.get());
-
-        osmdroidBasePathPref.set(osmdroidBasePath.getAbsolutePath());
-
-        File osmdroidTileCache = new File(osmdroidBasePath, "tiles");
-        //noinspection ResultOfMethodCallIgnored
-        osmdroidTileCache.mkdirs();
         Timber.d("Setting osmdroidBasePath to: %s", osmdroidBasePath.getAbsolutePath());
         configuration.setOsmdroidBasePath(osmdroidBasePath);
         Timber.d("Setting osmdroidTileCache to: %s", osmdroidTileCache.getAbsolutePath());
@@ -58,30 +49,25 @@ public class MapViewUtils {
 
         setMaxCacheSize(configuration);
 
+        // TODO Expiration! setExpirationExtendedDuration() OR setExpirationOverrideDuration()
+        // TODO Add option to adjust
+
         configuration.setMapViewHardwareAccelerated(true);
         configuration.setUserAgentValue(BuildConfig.APPLICATION_ID + "/"
                 + BuildConfig.VERSION_NAME + " " + org.osmdroid.library.BuildConfig.APPLICATION_ID
-                + "/" + org.osmdroid.library.BuildConfig.VERSION_NAME);
-
-        // remove this and use the default TileSourceFactory version again when
-        // https://github.com/osmdroid/osmdroid/issues/1050 has been resolved
-        final OnlineTileSourceBase MAPNIK = new XYTileSource("Mapnik",
-                0, 19, 256, ".png",
-                new String[] {
-                "http://a.tile.openstreetmap.org/",
-                "http://b.tile.openstreetmap.org/",
-                "http://c.tile.openstreetmap.org/" },"© OpenStreetMap contributors");
+                + "/" + org.osmdroid.library.BuildConfig.VERSION_NAME
+                + " (" + activity.getString(R.string.contact_email) + ")");
 
         MapTileProviderBasic mapnikTileProvider =
-                new MapTileProviderBasic(activity, MAPNIK);
+                new MapTileProviderBasic(activity.getApplicationContext(), WIKIMEDIA);
 
         MapView mapView = new MapView(activity, mapnikTileProvider);
-        mapView.setBuiltInZoomControls(true);
+        mapView.getZoomController().setVisibility(
+                CustomZoomButtonsController.Visibility.SHOW_AND_FADEOUT);
         mapView.setMultiTouchControls(true);
-        mapView.getController().setZoom(1);
+        mapView.getController().setZoom(1.0d);
         mapView.getController().setCenter(new GeoPoint(0.0d, 0.0d));
         mapView.setClickable(true);
-        mapView.setBuiltInZoomControls(true);
         mapView.setLayoutParams(new ViewGroup.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT,
                 ViewGroup.LayoutParams.MATCH_PARENT));
         mapView.setTilesScaledToDpi(true);
@@ -97,95 +83,6 @@ public class MapViewUtils {
         return mapView;
     }
 
-    @NonNull
-    private static File checkAndGetOsmdroidBasePathFile(Context context, String savedBasePath) {
-        Timber.d("Saved path is: %s", savedBasePath);
-        File savedBasePathFile = new File(savedBasePath);
-        //noinspection ResultOfMethodCallIgnored
-        savedBasePathFile.mkdirs();
-
-        boolean useSavedDir = isPathAvailableForWrite(context, savedBasePathFile);
-
-        if (!useSavedDir) {
-            Timber.d("Saved path unavailable, finding best storage");
-            savedBasePathFile = new File(getBestStorageLocation(context), "osmdroid");
-            //noinspection ResultOfMethodCallIgnored
-            savedBasePathFile.mkdirs();
-        }
-
-        Timber.d("Using path: %s", savedBasePathFile.getAbsolutePath());
-        return savedBasePathFile;
-    }
-
-    private static File getBestStorageLocation(Context context) {
-        Timber.d("Finding best storage location.");
-
-        ArrayList<File> storageDirs = new ArrayList<>(3);
-        storageDirs.add(context.getFilesDir());
-
-        File[] externalDirs = ContextCompat.getExternalFilesDirs(context, null);
-        for (File externalDir : externalDirs) {
-            // "Returned paths may be null if a storage device is unavailable."
-            if (externalDir == null) {
-                Timber.d("Storage location is null (=unavailable), skipping.");
-                continue;
-            }
-
-            String state = EnvironmentCompat.getStorageState(externalDir);
-            if (Environment.MEDIA_MOUNTED.equals(state)) {
-                storageDirs.add(externalDir);
-            }
-        }
-
-        File bestLocation = null;
-        long bestFreeSpace = 0;
-        for (File storageDir : storageDirs) {
-            long freeSpace;
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN_MR2) {
-                // gives more accurate information
-                freeSpace = new StatFs(storageDir.getAbsolutePath()).getAvailableBytes();
-            } else {
-                freeSpace = storageDir.getFreeSpace();
-            }
-            Timber.d("Found available storage: " + storageDir.getAbsolutePath()
-                    + ", free space: " + freeSpace);
-            if (freeSpace > bestFreeSpace) {
-                bestLocation = storageDir;
-                bestFreeSpace = freeSpace;
-            }
-        }
-
-        Timber.d("Determined best storage location: %s",
-                bestLocation != null ? bestLocation.getAbsolutePath() : "null");
-        return bestLocation;
-    }
-
-    private static boolean isPathAvailableForWrite(Context context, File path) {
-        boolean pathAvailable = false;
-        if (path.exists()) {
-            if (!isPathInInternalFilesDir(context, path)) {
-                String state = EnvironmentCompat.getStorageState(path);
-                pathAvailable = Environment.MEDIA_MOUNTED.equals(state);
-            } else {
-                // result of getFilesDir() is guaranteed to be writable
-                pathAvailable = true;
-            }
-        }
-        return pathAvailable;
-    }
-
-    private static boolean isPathInInternalFilesDir(Context context, File path) {
-        String canonicalPath;
-        String canonicalInternal;
-        try {
-            canonicalPath = path.getCanonicalPath();
-            canonicalInternal = context.getFilesDir().getCanonicalPath();
-        } catch (IOException e) {
-            return false;
-        }
-        return canonicalPath.startsWith(canonicalInternal);
-    }
-
     private static void setMaxCacheSize(IConfigurationProvider configuration) {
         // code adapted from osmdroid's DefaulConfigurationProvider.load()
         long cacheSize = 0;
@@ -196,10 +93,19 @@ public class MapViewUtils {
         }
 
         long freeSpace = configuration.getOsmdroidTileCache().getFreeSpace();
+        Timber.d("cacheSize: %d", cacheSize);
+        Timber.d("freeSpace: %d", freeSpace);
+        Timber.d("getTileFileSystemCacheMaxBytes(): %d",
+                configuration.getTileFileSystemCacheMaxBytes());
 
         if (configuration.getTileFileSystemCacheMaxBytes() > (freeSpace + cacheSize)) {
             configuration.setTileFileSystemCacheMaxBytes((long)((freeSpace + cacheSize) * 0.95));
             configuration.setTileFileSystemCacheTrimBytes((long)((freeSpace + cacheSize) * 0.90));
         }
+
+        Timber.d("getTileFileSystemCacheMaxBytes(): %d",
+                configuration.getTileFileSystemCacheMaxBytes());
+        Timber.d("getTileFileSystemCacheTrimBytes(): %d",
+                configuration.getTileFileSystemCacheTrimBytes());
     }
 }
